[refactor] Move tests out of the pages folder
NextJS does not support having tests along with source files.
This moves the tests into a __tests__ folder outside.

This fixes npm run build errors that were trying to include the tests in
the production build.
